mamute_scrappers/requirements.txt
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0

